+++ gradle-plugin/src/main/kotlin/dev/zacsweers/metro/gradle/MetroGradleSubplugin.kt
@@ -16,7 +16,9 @@ import org.jetbrains.kotlin.gradle.plugin.KotlinPlatformType
 import org.jetbrains.kotlin.gradle.plugin.SubpluginArtifact
 import org.jetbrains.kotlin.gradle.plugin.SubpluginOption
 import org.jetbrains.kotlin.gradle.plugin.getKotlinPluginVersion
+import org.jetbrains.kotlin.gradle.plugin.kotlinToolingVersion
 import org.jetbrains.kotlin.gradle.tasks.KotlinCompilationTask
+import org.jetbrains.kotlin.tooling.core.KotlinToolingVersion
 
 public class MetroGradleSubplugin : KotlinCompilerPluginSupportPlugin {
   private companion object {
@@ -24,6 +26,8 @@ public class MetroGradleSubplugin : KotlinCompilerPluginSupportPlugin {
       lazy(LazyThreadSafetyMode.NONE) {
         KotlinVersion.fromVersion(BASE_KOTLIN_VERSION.substringBeforeLast('.'))
       }
+
+    val kotlin230 = KotlinToolingVersion(2, 3, 0, null)
   }
 
   override fun apply(target: Project) {
@@ -141,6 +145,17 @@ public class MetroGradleSubplugin : KotlinCompilerPluginSupportPlugin {
         KotlinPlatformType.wasm -> false
       }
 
+    val kotlinVersion = project.kotlinToolingVersion
+    val orderComposePlugin = kotlinVersion >= kotlin230
+    kotlinCompilation.compileTaskProvider.configure { task ->
+      if (orderComposePlugin) {
+        // Order before compose-compiler
+        task.compilerOptions.freeCompilerArgs.add(
+          "-Xcompiler-plugin-order=${PLUGIN_ID}>androidx.compose.compiler.plugins.kotlin"
+        )
+      }
+    }
+
     // Ensure that the languageVersion is 2.x
     kotlinCompilation.compileTaskProvider.configure { task ->
       task.doFirst { innerTask ->
@@ -257,6 +272,8 @@ public class MetroGradleSubplugin : KotlinCompilerPluginSupportPlugin {
           )
         )
         add(lazyOption("contributes-as-inject", extension.contributesAsInject))
+        // Track whether we ordered the plugin before compose-compiler
+        add(SubpluginOption("plugin-order-set", orderComposePlugin.toString()))
         reportsDir.orNull
           ?.let { FilesSubpluginOption("reports-destination", listOf(it.asFile)) }
           ?.let(::add)